+++ client/examples/transaction_signing.rs
@@ -16,7 +16,8 @@ fn main() {
 async fn go() -> Result<(), Error> {
     let alice = ed25519::Pair::from_string("//Alice", None).unwrap();
     let bob = ed25519::Pair::from_string("//Bob", None).unwrap();
-    let client = Client::create().compat().await?;
+    let node_host = url::Host::parse("127.0.0.1").unwrap();
+    let client = Client::create(node_host).compat().await?;
 
     let account_nonce = client.account_nonce(&alice.public()).compat().await?;
     let transfer_tx = Transaction::new_signed(